backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.3
PyJWT==2.9.0
pyarrow==22.0.0
pandas==2.3.3
numpy==2.3.4
boto3==1.40.61
pytest==8.3.3
